third_party/blink/web_tests/platform/mac-mac10.14/external/wpt/paint-timing/fcp-only/fcp-ensure-update-the-rendering-step-expected.txt

This is a testharness.js-based test.
FAIL The first-contentful-paint timestamp should be same as the last RAF assert_approx_equals: One of the past requestAnimationFrame should have the same timestamp as paint entry expected 49.369 +/- 1 but got 67.49999999919964

assert_approx_equals: One of the past requestAnimationFrame should have the same timestamp as paint entry expected 41.86 +/- 1 but got 67.49999999919964
Harness: the test ran to completion.
